pulling shared functions to reorganize files
