+++ app/routes/app.experiments.new.jsx
@@ -18,6 +18,9 @@ import { useFetcher, redirect, useLoaderData } from "react-router";
 import { useState, useEffect, useRef } from "react";
 import db from "../db.server";
 import { ExperimentStatus } from "@prisma/client";
+import { TimeSelect } from "../utils/timeSelect";
+import { validateStartIsInFuture } from "../utils/validateStartIsInFuture";
+import { validateEndIsAfterStart } from "../utils/validateEndIsAfterStart";
 
 // Server side code
 export const action = async ({ request }) => {
@@ -278,247 +281,6 @@ export const loader = async ({ request }) => {
 
 //--------------------------- client side ----------------------------------------
 
-function TimeSelect({
-  id = "selectTime",
-  label = "Select time",
-  value,
-  onChange,
-  error,
-  invalidMessage = 'Enter a time like "1:30 PM" or "13:30"',
-}) {
-  // controlled display value (human readable like "1:30 PM")
-  const times = [];
-  const popoverId = `${id}-popover`;
-  // build times list once
-  for (let h = 0; h < 24; h++) {
-    for (let m = 0; m < 60; m += 30) {
-      const hour24 = h.toString().padStart(2, "0");
-      const minute = m.toString().padStart(2, "0");
-      const value24 = `${hour24}:${minute}`;
-      const suffix = h >= 12 ? "PM" : "AM";
-      const hour12 = ((h + 11) % 12) + 1;
-      const label12 = `${hour12}:${minute} ${suffix}`;
-      times.push({ value: value24, label: label12 });
-    }
-  }
-
-  const labelFor = (hhmm) => {
-    if (!hhmm) return "";
-    const hit = times.find((t) => t.value === hhmm);
-    if (hit) return hit.label;
-    const [H, M] = hhmm.split(":").map((n) => parseInt(n, 10));
-    const am = H < 12;
-    const h12 = ((H + 11) % 12) + 1;
-    return `${h12}:${String(M).padStart(2, "0")} ${am ? "AM" : "PM"}`;
-  };
-
-  // local display state so we can show the friendly label while remaining controlled
-  const [display, setDisplay] = useState(value ? labelFor(value) : "");
-  useEffect(() => {
-    
-    // sync whenever parent value changes (including when validation sets an error)
-    setDisplay(value ? labelFor(value) : "");
-  }, [value]);
-
-  const openPopover = (el) =>
-    el?.querySelector(`#${popoverId}Trigger`)?.click();
-
-  const commitFromField = (raw) => {
-    const parsed = parseUserTime(raw);
-    if (!parsed) {
-      // notify parent by passing null / empty so parent can set error string
-      onChange("");
-      return;
-    }
-    onChange(parsed);
-    setDisplay(labelFor(parsed));
-  };
-
-  return (
-    <div>
-      {/* This section is what will visually display when the function is called */}
-      <s-text-field
-        label={label}
-        id={`${id}-input`}
-        icon="clock"
-        value={display}
-        placeholder="Choose a time"
-        error={error}
-        onFocus={(e) => openPopover(e.currentTarget.parentElement)}
-        onClick={(e) => openPopover(e.currentTarget.parentElement)}
-        onInput={(e) => setDisplay(e.currentTarget.value)}
-        onBlur={(e) => commitFromField(e.currentTarget.value)}
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            e.preventDefault();
-            commitFromField(e.currentTarget.value);
-          }
-        }}
-      >
-        <s-button
-          slot="accessory"
-          variant="tertiary"
-          disclosure="down"
-          commandFor={popoverId}
-          icon="chevron-down"
-          accessibilityLabel="Select time"
-        />
-      </s-text-field>
-
-      {/* This is the popover styling and the button population */}
-      <s-popover id={popoverId} maxBlockSize="200px">
-        <s-stack direction="block">
-          {times.map((t) => (
-            <s-button
-              key={t.value}
-              fullWidth
-              variant="tertiary"
-              commandFor={popoverId}
-              onClick={() => {
-                onChange(t.value);
-                setDisplay(labelFor(t.value));
-              }}
-            >
-              {t.label}
-            </s-button>
-          ))}
-        </s-stack>
-      </s-popover>
-    </div>
-  );
-}
-
-//This function cleans and parses the user input, we only care about numbers and :, everything else is scrubbed
-function parseUserTime(input) {
-  if (!input) return "";
-  let s = String(input)
-    .trim()
-    .toLowerCase()
-    .replace(/\s+/g, "")
-    .replace(/\./g, "");
-  if (s === "noon") return "12:00";
-  if (s === "midnight") return "00:00";
-  let ampm = null;
-  if (s.endsWith("am")) {
-    ampm = "am";
-    s = s.slice(0, -2);
-  } else if (s.endsWith("pm")) {
-    ampm = "pm";
-    s = s.slice(0, -2);
-  }
-  s = s.replace(/[^0-9:]/g, "");
-  let hh = 0,
-    mm = 0;
-  if (s.includes(":")) {
-    const [hStr, mStr = "0"] = s.split(":");
-    if (!/^\d+$/.test(hStr) || !/^\d+$/.test(mStr)) return null;
-    hh = parseInt(hStr, 10);
-    mm = parseInt(mStr.padEnd(2, "0").slice(0, 2), 10);
-  } else {
-    if (!/^\d+$/.test(s)) return null;
-    if (s.length <= 2) {
-      hh = parseInt(s, 10);
-      mm = 0;
-    } else if (s.length === 3) {
-      hh = parseInt(s.slice(0, 1), 10);
-      mm = parseInt(s.slice(1), 10);
-    } else {
-      hh = parseInt(s.slice(0, -2), 10);
-      mm = parseInt(s.slice(-2), 10);
-    }
-  }
-
-  //error handling for if minutes are out of bounds
-  if (isNaN(hh) || isNaN(mm) || mm < 0 || mm > 59) return null;
-
-  //error handling for if user types in am/pm to check that hours are within bounds
-  if (ampm) {
-    if (hh < 1 || hh > 12) return null;
-    if (ampm === "am") {
-      if (hh === 12) hh = 0;
-    } else {
-      if (hh !== 12) hh += 12;
-    }
-  } else {
-    if (hh < 0 || hh > 23) return null;
-  }
-  //This is what we care about most, returns a string in 24hr format with hh:mm
-  return `${String(hh).padStart(2, "0")}:${String(mm).padStart(2, "0")}`;
-}
-
-function validateStartIsInFuture(startDateStr, startTimeStr = "00:00") {
-  let dateError = "";
-  let timeError = "";
-
-  if (!startDateStr) {
-    return { dateError, timeError };
-  }
-
-  const today = new Date();
-  today.setHours(0, 0, 0, 0);
-  const selectedDate = new Date(`${startDateStr}T00:00:00`);
-
-  if (selectedDate < today) {
-    dateError = "Start date cannot be in the past";
-    return { dateError, timeError: "" };
-  }
-
-  const isToday = selectedDate.getTime() === today.getTime();
-
-  if (isToday) {
-    const startDateTime = new Date(
-      `${startDateStr}T${startTimeStr || "00:00"}`,
-    );
-    const now = new Date(); // The *actual* current time
-
-    if (startDateTime <= now) {
-      timeError = "Start time must be in the future";
-    }
-  }
-  return { dateError, timeError };
-}
-
-function validateEndIsAfterStart(
-  startDateStr,
-  startTimeStr = "00:00",
-  endDateStr,
-  endTimeStr,
-) {
-  // return both a date-level error and a time-level error so UI can show the right one
-  let dateError = "";
-  let timeError = "";
-
-  if (!startDateStr || !endDateStr) {
-    return { dateError, timeError };
-  }
-
-  const effectiveEndTime = endTimeStr || "23:59";
-  const startDateTime = new Date(`${startDateStr}T${startTimeStr || "00:00"}`);
-  const endDateTime = new Date(`${endDateStr}T${effectiveEndTime}`);
-
-  if (endDateTime <= startDateTime) {
-    const startDateOnly = new Date(`${startDateStr}T00:00:00`);
-    const endDateOnly = new Date(`${endDateStr}T00:00:00`);
-    // if end date precedes start date -> show error on date
-    if (endDateOnly.getTime() < startDateOnly.getTime()) {
-      dateError = "End date must be after the start date";
-    } else {
-      // same day but time invalid -> show error on time
-      timeError = "End time must be after the start time";
-    }
-  }
-  return { dateError, timeError };
-}
-
-// convert a local date (YYYY-MM-DD) and local time (HH:MM) into a UTC ISO string
-function localDateTimeToISOString(dateStr, timeStr = "00:00") {
-  if (!dateStr) return "";
-  const [y, m, d] = dateStr.split("-").map(Number);
-  const [hh = 0, mm = 0] = (timeStr || "00:00").split(":").map(Number);
-  // construct a local Date from components (guaranteed local interpretation)
-  const local = new Date(y, m - 1, d, hh || 0, mm || 0, 0, 0);
-  return local.toISOString(); // canonical UTC instant
-}
 
 export default function CreateExperiment() {
   //fetcher stores the data in the fields into a form that can be retrieved
